src/requirements.txt
python-telegram-bot==21.3
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1